Optimize for file size and performance
The subGroup way of dealing with multiple selectors has been removed and
replaced with one where a list of unique qualified names from the
multiple selectors is created during parsing, and if there exists more
than one unique name, all DOM elements are selected using "*" and are
pre-filtered for only those qualified names before running the
selectors. This removes any need to check for uniqueness and to sort the
result.